BuildClientLogger moved to bep package
