+++ src/main/java/org/jetbrains/bsp/bazel/server/bep/loggers/BUILD
@@ -1,7 +1,7 @@
 load("@rules_java//java:defs.bzl", "java_library")
 
 java_library(
-    name = "logger",
+    name = "loggers",
     srcs = glob(["*.java"]),
     visibility = ["//visibility:public"],
     deps = [
