+++ src/main/java/org/jetbrains/bsp/bazel/server/bsp/services/BUILD
@@ -13,7 +13,6 @@ java_library(
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp/resolvers",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp/utils",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/logger",
         "@io_bazel//src/main/protobuf:build_java_proto",
         "@io_bazel//third_party/grpc:grpc-jar",
         "@maven//:ch_epfl_scala_bsp4j",
